At the moment, it's possible to create a review request against any changeset (with the perforce backend).  If one person creates a review against another person's changeset number, that will then prevent the owning user from creating a request.
Since perforce changelists include usernames, it's not hard to
parse this out before we create anything in the database and
check it against the logged-in user.

Part of this change also fixes a bug in the "New Review Request"
form where the selected repository was not being marked as such
if there was an error.

This also shows a nice error if the uploaded file isn't a diff.

Reviewed at http://reviewboard.chipx86.com/r/35/