This is a generic solution to the question of adding default reviewers. It's generic enough to be useful both for installations like reviews.r-b.o where there's a single group for all reviews, and the extraordinarily complex case we've got at VMware, where different groups own different parts of a single huge codebase.
There's a new model which basically has a regular expression and a list
of default people/groups.  When a new diff is created, it iterates through
the list of filenames in the diff and checks them against each of the
DefaultReviewer regexps.  If one matches, those defaults are added to the
draft's list of target reviewers.

Reviewed at http://reviews.review-board.org/r/169/