Patch from Xavier Duret to make post-review work on machine configured in non english locale.
The client side "post-review" script does not work when used on a machine
configured as a non English locale. This is due to the fact that the tool parses
the output of the subversion commands which is not necessarily written in
English. This patch makes sure the subversion command is called with the proper
locale. This patch affect only subversion but it is possible that other revision
control tools also need it.

Testing done:
Tested on a machine with locale "en_US.UTF-8" and a machine with locale "es_ES.UTF-8".

Reviewed at http://reviews.review-board.org/r/419/