Turn off translation for now.
While we get pages partially translated because of django's built-in strings, we
don't actually have any translations yet. The result of this is that when a
user has their language set, they get a partially translated page, which is
confusing. Once people start translating, we can turn this back on.

Fixes bug 500.