This introduces improved support for fastcgi in lighttpd and Apache, as well as improved lighttpd support in general.
Both lighttpd and Apache will start our newly generated fastcgi script on server startup. This removes the need for people to provide their own init scripts for starting fastcgi. That option is still available to people, of course, since users can always manually change their server config, but an out-of-the-box install of Review Board using fastcgi will now start up without additional work by the user.

This also improves lighttpd by adding support for compression (as well as caching of compressed files -- static text files only), adding a 1 year cache time of media files, and by defaulting to including all configuration inside a vhost.

As part of this, rb-site generates a "tmp" directory inside the site configuration directory that socket files, cached media, etc. can be written to. There's nothing in place yet for helping the user clear this directory, but we're not doing much with it at this point, so it shouldn't get too full. In the future we can use this for more things.

Reviewed at http://reviews.review-board.org/r/719/