Add a new administration guide, users guide, and FAQ.
This is the culmination of nearly three months of effort. We have new
docs! From now on, documentation will be maintained within the
reviewboard source tree, and viewed on the website.

They're searchable, thanks to Google Custom Search.
They're fully fleshed out with screenshots.
They're easy to maintain, using ReST and Sphinx.
They also answer many questions that our old docs did not. Questions
that we've answered time and time again on the mailing list.

The current doc set contains everything we had on our wiki, and much,
much more.