Fix the last_activity_timestamp query to work with MySQL.
The query requires wrapping ORDER BY with something like MAX() when
using PostgreSQL, and it works fine with sqlite as well, but MySQL has
problems with it where it will return an empty set when used against a
review request with no reviews, so we need to use it conditionally.

It's an ugly hack.  Post-1.0, we really should just store a timestamp
along with the ReviewRequest containing the last activity. The problem
is that we end up bumping last_updated during migrations, which is bad,
so we need to somehow be clever with this and only update last_updated
when we absolutely have to.