Make our unit tests relocatable.
The unit tests assumed the tests were being run from a specific
directory, and would break if not run from that directory. This was
usually okay, but it's not clean, and will break with some upcoming
directory restructuring work, so this change fixes the tests to work no
matter where the tests are being run from.