Chain up API calls to the server.
We were calling API functions in parallel, which in some cases could
cause breakages on the server when the server was under load. For
example, clicking Save on a comment when another comment was still being
saved.

We now chain up rbApiCalls that POST to the server so that only one API
request is happening at a time. This goes a long way to keep from having
broken reviews on the server when things are stalled.

Reviewed at http://reviews.review-board.org/r/892/