Better check for needed executables in the PATH.
This change reduces much of the code in Review Board for checking for the
existance of certain executables in the PATH. We now use
djblets.util.filesystem.is_exe_in_path instead, which both cleans/centralizes
our code and checks for the exact binary name we care about on the platform.

Another benefit is that we now use the proper path separator, instead of
hoping $IFS is set.

Reviewed at http://reviews.review-board.org/r/918/
Fixes bug #1180