Allow post-commit to recognize reviewboard URLs without a trailing slash.
This is a trivial change for the post-commit hook that lets it still work if the
URL in the commit message lacks a trailing slash.  I came across this when some
of my users would put something like this in the commit message:

    Review: http://REVIEWBOARD_URL/r/123

instead of

    Review: http://REVIEWBOARD_URL/r/123/

Testing done:
This change has been in use on my install for a while and is working as
expected.

Patch from Russell Bryant
Reviewed at http://reviews.review-board.org/r/759/