Implemented the detection of whitespace-only changes in diff.
On the client side, added a button to the diff viewer to toggle between whitespace visible and highlighted, or hidden and regular. On the server, added metadata on the line/chunk/file level, so that the interface can use it to toggle.

Toggled chunks are not collapsed, just no longer highlighted. For mixed chunks, the lines that contain only whitespace are set to the regular background, making the visualization better.

Testing done:
- Ran all unit tests currently available, without a breakage.
- Applied patch in a production system, with no breakage of existing reviews.
- Tested on IE 7, Chrome, Firefox 3.5 and 3.0 and Safari 4.0.

Reviewed at http://reviews.review-board.org/r/899/