Add raw file URL support to the Repository admin page and Git.
This introduces a new column, raw_file_url, to Repository, and shows it in the
admin UI for SCMTools that support it. This URL provides a way for certain
SCMTools (right now, Git only) to access files over HTTP, instead of checking
the file out through the SCM or accessing it locally (in the case of Git).

The raw file URL acts as a URL with fields to be substituted. If <revision>
exists in the URL, it will be replaced with the revision of the file. If
<filename> exists, it will be replaced with the filename.

This also adds a GitHub hosting service and bug tracker to the Repository
admin. Using this, it's trivial to add any GitHub project without knowing the
raw file URL mask.

Reviewed at http://reviews.review-board.org/r/977/