Fix sending e-mails from publishing drafts.
There were some issues with sending e-mails when publishing drafts. The
review_request_published signal was being emitted with the wrong arguments.
This is now fixed. The side-effect is that the notification was emitted
twice, once by ReviewRequest.publish, once by ReviewRequestDraft.publish. We
now have a send_notification flag to ReviewRequestDraft.publish that
ReviewRequest.publish sets to False, to prevent this.

Also, the webapi unit tests weren't initializing Review Board, which meant
that no signals for e-mail would be published, and the tests would fail.