Only look for AD group membership when actually needed
Patch by Nikolai Prokoschenko. This changes the Active Directory functionality a
bit: now the group membership is only checked when an actual restriction on
groups is active. If not, not groups are retrieved.

This can also be considered a workaround for issue 1267

Testing done: Patched and implemented at my workplace.

Reviewed at http://reviews.review-board.org/r/1005/
        and http://reviews.review-board.org/r/956/