Make it clear that Review Board requires JavaScript.
If someone tries to use Review Board with a browser that doesn't support
JavaScript, or has JS turned off, random bits of important functionality will
just not work. This is a pretty crappy failure mode.

This change adds an "important" box at the top of the page inside a <noscript>
tag, explaining that JS needs to be turned on.