Improved admin site usability by making subnavigation look and behave like tabs.
When configuring ReviewBoard it was really easy to loose track on which tab you were, since there was no visual cue to that.
This patch makes really easy for the user to recognize where they are by adding a lighter background on the front tab.
As asked by Christian during review, it is done using a templatetag and url reverse().

Reviewed at http://reviews.review-board.org/r/1029/