Relax Admin UI restrictions on the Bugs Closed field.
The Bugs Closed field was set as a CommaSeparatedIntegerField, even
though we've relaxed other restrictions to allow for alphanumeric
bug numbers. This field is now a CharField, meaning that it can
technically contain anything, and we can restrict it on other
layers.

This doesn't require a database schema evolution, as
CommaSeparatedIntegerField is just a CharField with some extra
validation in the admin UI.

Fixes bug #1149.