Fix a 500 error when attempting to submit as a user without valid permissions.
We had a nasty 500 error due to attempting to submit a change to a review
request as another user without having valid permissions to do so. This was
due to returning a WebAPIResponseError in _prepare_draft, but never checking
if that was the result of the function. So, we'd crash when attempting to
treat that resulting WebAPIResponseError as a ReviewRequestDraft.

_prepare_draft now raises PermissionDenied, which we catch on each call
site and then return the appropriate response.

Added a unit test to check for this.

Fixes bug #1067