Make the "Save" or "Publish" work a second time in the review dialog.
There was a bug where the "Save" and "Publish" buttons in the review dialog
wouldn't work after a second time. This was due to our function queue for
handling the saving being marked as in-progress. We now clear this
function queue before saving/publishing, allowing the button to work multiple
times without reloading the page.

Fixes bug #1292