Filter out inactive users from the dashboard.
We've always filtered out inactive users from the Submitters and Group
Members lists, but never from the dashboard or All Review Request pages.
This was an oversight.

When a user leaves a company/organization and is marked inactive, it's annoying
to see a bunch of stale review requests hanging around taking up space, and
it's up to moderators to close them all out. Now administrators can simply
mark the user as Inactive and their review requests will be hidden from
view. They're still accessible if you have the link, however, so nothing is
really gone. They're just not taking up the space anymore.

Fixes bug #1290