Relax the server URL format in the General Settings page.
We were using a URLField for the server URL in the General Settings
page. This pretty much makes sense, except that Django's URLField
assumes a standard public domain name, "localhost", or IP address.
Some of our users, though, use simply "reviewboard" or "reviewboard.internal"
or some variation, which URLField doesn't allow.

Since we don't do any checks during rb-site install anyway (which would
skip the restrictions on this field), and this is a field that administrators
would know better than us, we're moving from a URLField to a plain CharField.
This gives the greatest flexibility, and we can always add checks on from
there if we need to.

Fixes bug #1223.