Increase the FileDiff filename limit to 1024 characters.
The filename limit in FileDiffs were previously 256 characters, which was
fine for Subversion and other systems generally, but it was possible to hit
this boundary and get some confusing errors. This is far more common with
ClearCase, where filenames can get quite large.

The maximum filename size in ClearCase is 1024 characters, which seems like
a pretty decent size. This change bumps up the max size to 1024 characters
and adds a unit test to make sure this does in fact work.

Fixes bug #751
Reviewed at http://reviews.review-board.org/r/1034/