Show the draft screenshot caption for draft review requests.
We had a bug where the draft caption for a screenshot wasn't shown when
the user modified it and had the draft open. This was because we were
referencing only the published caption. Now we check if we have a draft and,
if so, show the draft caption.

Fixes bug #1257