Fix expanding some collapsed diffs in interfilediffs.
There was a bad assumption made when expanding files in interfilediffs
when the file containing the chunk to expand didn't actually have an
associated interfilediff. We were always using the global interfilediff
revision in the diff viewer, instead of only using it for the diffs that
needed them. We now pass in the proper revision and interdiff revision for
each expand link.

Expanding diffs should always work now for all types of diffs.

Fixes bug #1285