Add Amazon S3-based screenshot storage.
This builds upon my djblets changes to allow people to use django-storages'
Amazon S3 storage backend. With this change, admins enter their Amazon AWS
access key and secret key, and an S3 bucket name. Uploaded screenshots get saved
to S3, and so do the according thumbnails.

This change is almost entirely admin site/siteconfig updates for the UI to set
the various settings keys that django and django-storages use to select the
storage backend and configure it. The only change to non-admin code is to change
the screenshot serializer to use the basename of the filename instead of the
basename of the path, since not all storage backends implement the path
property.

Reviewed at http://reviews.review-board.org/r/1080/