Move the configuration templates into reviewboard/cmdline/conf/.
We moved the codebase into the reviewboard/ directory, but didn't move the
contrib/ directory or anything into it, meaning that they weren't bundled
in the Python packages. Because of this, rb-site couldn't fetch the templates
for the server configuration files.

Now the configuration has been moved into reviewboard/cmdline/conf/. This
fixes rb-site and allows new sites to be installed again.