Add improved autocomplete of groups and people.
This change introduces a much improved autocomplete for groups and people.
In the case of people, we now search the username, first name, and last name.
In the case of groups, we now search the group name and the display name.

In all cases, the full name (for people) and display name (for groups) are
displayed next to the username/group name, right-aligned.

Users no longer have to guess at a person's username to add them.

Fixes bug #598
Reviewed at http://reviews.review-board.org/r/1081/