Change the semantics of a Review Request's last activity information.
This changes what ReviewRequest.get_last_activity covers. Previously, it
could return private updates, such as the update to a review request's draft.
Really, though, this is never what we want. It was causing update bubbles
when a user was editing his own review request, and that's never really
correct. The only other place we ever used this was for generating ETags, but
the draft can be considered separately from this time anyway.

Now, get_last_activity will only return information on publicly visible
objects. This fixes the excess update bubble issue, and makes it a lot
more clear what callers can expect from this function.