Strip whitespace from usernames before creating users.
We have cases where a username with leading or trailing whitespace could
be used for looking up and creating a user. While not a problem with
standard auth, this is a problem with LDAP and some other auth backends.
We now strip the whitespace when authenticating and when looking up/creating
a User.

As an extra precaution, we also now strip when setting the reviewers, before
calling into the auth backends.

Fixes bug #1309