Store the interfilediff of a comment on its reply.
When replying to a comment, we were transferring the filediff to the reply,
but not the interfilediff. This meant that replies wouldn't show up in the
diff viewer in the discussion area and links in the e-mail were wrong.

We now simply set the interfilediff of the reply to the comment on creation.

Fixes bug #765