Don't show a 500 error when failing to load a repository on the New Review Request page.
If there was an error in loading a SCMTool for a repository (for example,
an ImportError), then the New Review Request page would fail to load, showing
a 500 error instead. We now catch all exceptions in loading a review request
and then log it to the log file and remove it from the list shown.