Include the diff revision number in the cache key.
We had a stale cache issue with the diff fragments. If a user uploaded a diff
to a draft of a new review request, viewed the diff, and then published the
diff, then the version they viewed would be served up to new users. However,
the generated HTML made use of the revision number in the Expand link, which
would be 0 due to the draft diff. We now include the revision in the cache key
so that we don't hit this issue.

Fixes bug #1334