Don't break on entering bugs when a bug tracker URL is empty.
When entering bug numbers into a review request with a repository that doesn't
have a bug tracker URL set, we were attempting to display the resulting array
of bug numbers as a string, instead of transforming it into a string. We now
transform it into a string before displaying it.

Fixes bug #1195