Fix screenshot comment issues with Google Chrome.
On Google Chrome, the image's margin-left needs to be added to the position
of the image in order to get the correct location of the image to use. On
other browsers, the margin-left will appear as 0 (as we'd expect), so we just
now add this value to the position we're going to set for the screenshot
comment area.

Fixes bug #1050