Change the layout of the header bar.
The header bar layout had various issues on different browsers, since we
essentially tried to fit the account navbar and the Review Board logo
and version info on the same horizontal space.

We now display the account navbar at the top with a border below it, like
many other sites have it. The Review Board logo and version info is now below
this.

This causes a bit of a structural change to the page. Now everything below
the header bar is inside a page-container div with some padding. The base
body has no margin or padding anymore.

Tested on Firefox, Opera, IE, and Chrome.