Add the ability to set a repository's visibility.
When an organization moves to a new repository, they're supposed to create
a new Repository entry instead of changing an existing one so that they
don't break existing review requests. However, the old repositories must
stick around, and are seen in the New Review Request page and in the
Default Reviewer lists.

This change adds a 'visible' field to Repository to indicate whether or not
it should show up in user-visible lists. We use this to filter on the
New Review Request and Default Reviewer pages.

This does not prevent admins from seeing the repository on pages such as the
Review Request database entry, since it's still important to see that, and
doesn't prevent post-review from working, since that means the repository
is still presumably around. It's just a way to help prevent people from
selecting old repositories when they didn't mean to.

Reviewed at http://reviews.review-board.org/r/1153/