Fix a TypeError when sorting large bug numbers.
When entering multiple long numbers in the bug list of a review request,
posting the request resulted in a TypeError. This is due to bug numbers
of certain really long lengths being longs and not ints, and our math
for numerical sorting therefore returning a long instead of an int.
list.sort expects that the function passed to cmp will return an int, so
now we cast the result of the operation to an int.

Fixes bug #1352
Patch by Thilo-Alexander Ginkel
Reviewed at http://reviews.review-board.org/r/1168/