Make sure to return a unicode in ReviewRequest.__unicode__.
If a review request didn't have a summary, ReviewRequest.__unicode__ would
attempt to return a translated "(no summary)". However, it did that with
ugettext_lazy, which returned a proxy object which wasn't always acceptable
as a string. We now make sure to return this as an actual unicode string.

Added a test case to check for this.

Fixes bug #1346