Remember the previously stored base diff path.
Users who use the Upload Diff form instead of post-review against Subversion
repositories (and others) have to type a base diff path every time, which
can be annoying. This change stores the base diff path on uploads and re-uses
it as the default for the field the next time the user goes to upload a
diff.

Fixes bug #365
Reviewed at http://reviews.reviewboard.org/r/1172/