Several improvements to the New Review Request page.
This changes the labels and help text on the New Review Request page to
be the same as the Upload Diff dialog, and to increase the width of the
fields.

It also adds some informational text above the New Review Request box that
recommends using post-review, very briefly describes it, and points to the
documentation. In the future, this will be updated to point to a page linking
to first and third party tools, such as IDE plugins.

Fixes bug #1355