Fix retrieving file contents in Bazaar when the file was moved in a previous revision.
Due to the order in which we performed a couple operations, we would fail at
retrieving a file's contents if the file was moved in a previous revision.
We now order things correctly so as to prevent this.

Patch by Kay Röpke
Fixes bug #1393
Reviewed at http://reviews.reviewboard.org/r/1209/