Fix an issue with displaying comment flags on consecutive lines.
If two comments were placed on two consecutive lines, we would only show the
first comment. This appears to be due to an off-by-one when factoring in
the table headers *and* a starting row. Since the starting row includes the
table headers, we don't want to factor in the headers a second time. We now
make sure we don't.

Patch by Laurent Nicolas
Fixes bug #1377
Reviewed at http://reviews.reviewboard.org/r/1189/