Move raw patch logic out of view and into diffparser.
Moves "create diff/patch" logic out of the view and into the diffparser. This
way the view just deals with "strings" rather than having to create diff
fragments.

Moving the logic out of the view also means that the raw patch can be exposed
in other places, e.g. we could expose the diff in an email (NOTE if this was
done would need to be configurable so that normal users don't suffer the
overhead of generating raw patches, e.g. lazy/deferred evaluation or a config
option).

Patch by Chris Clark
Reviewed at http://reviews.reviewboard.org/r/1179/