Fix uploading of diffs through the web API.
A recent change to fix creating review requests through the New Review Request
page actually broke several things. It broke the web API, which was no longer
causing the form data to be passed through properly, and it failed to catch
the new explicitly named exceptions, as they were never imported.

We now pass in the data in all cases properly, and we also have a new unit
test for the New Review Request page to ensure that it doesn't break again.