Detect and display moved parts in a file.
This introduces move detection to the diff viewer. The diff viewer now
attempts to match up insert blocks with corresponding delete blocks in order
to detect if a group of code has moved elsewhere in the file.

If it has moved, we display little tags alongside the code for jumping to the
source/destination of the line. The line will briefly highlight yellow and
fade out.

This is a Google Summer of Code 2009 project by Eduardo Felipe Castegnaro.
More details can be found on the review request.

Patch by Eduardo Felipe Castegnaro
Reviewed at http://reviews.reviewboard.org/r/926/