Improve error handling when uploading a diff and when using Git.
If a diff upload failed for whatever reason, there was no log information
to show what went wrong. We now log the exception trace to the log file.

If Git fails to access a repository, we log the error message as well. This
is perhaps less useful generally, but there's no reason not to have it.