Some fixes for diff fragment headers.
We were showing the <tbody> tags for diff fragment headers even when we didn't
have any headers. We now only show these if we do have them.

Also, our default sizes for the fragments while loading now assumes a header.
This won't always be true, but for most source code, it will probably be
true more often than not. Unfortunately, we don't have a way today of flagging
this in a comment in order to be more intelligent about it.