Fix issue with "comment flag not visible in multi-page collapsed diffs"
Patch from Laurent Nicolas. Some miscalculation in rb javascript may cause a
comment flag not to be shown in the diff view.  The comment is still present in
the DB and is shown in the reviews view.  Also improved the code around
guessRowNum, so it may be a bit speedier if the commented line is close to the
row being examined.

Testing done:
Performed UT and Integration Testing using firebug.
Verified that the initial issue reported by end user is fixed for their diffs.

Fixes bug 1165