Improve the rb-site installer UI.
This makes several UI improvements for the rb-site installer.

The Review Board logo is now shown in the top-left of the installer pages,
backed by a white (usually -- this is dependent on the theme) box containing
the page title.

There are separators between the title box and the content and the button box.

The window now has an icon (the favicon for 16x16 and the larger logo for
larger sizes).

The "Forward" button has been renamed "Next."

Reviewed at http://reviews.reviewboard.org/r/1444/