Add the logo to the admin UI, which links to the main site.
Added the Review Board logo to the admin UI. Clicking on it jumps to the
main site.

Down the road it'd be nice to have a more explicit "Back to site" link, but
I want to change the design of the admin UI before doing this, as it adds
more navigation clutter otherwise.

Fixes bug #1299