Improve several forms in the database browser.
This change enhances several forms in the database browser.

The organization is cleaned up in some forms, particularly the Review Request
and draft forms. The fields are in clear sections, arranged well, and some
of the more advanced state is hidden in a collapsed fieldset. Fields that
would previously load in potentially hundreds of thousands of entries now
use a raw ID field.

Some forms (such as the diff-related forms) inline their children, making it
a little easier to see useful details.

Many of the object display pages now have proper ordering and allow the
user to search through the objects.

This also adds some actions to the Review Request objects for closing and
reopening review requests.

Reviewed at http://reviews.reviewboard.org/r/1446/