Add support for custom site-specific management commands.
This change introduces the ability to write custom management commands specific
to a site installation.

This uses standard Django management commands. They can be placed in
the site's commands/ directory. `rb-site manage` will look for Python files
in this directory and add them to the known list of management commands.

Today, adding them to the list is pretty hacky, but there aren't any other
options. We will probably want to request the ability to cleanly register
new commands in Django. It works though, and we can update it if needed
down the road.

Reviewed at http://reviews.reviewboard.org/r/1447/