Fix rbApiCall with the new HTTP status codes.
The new HTTP status codes caused our API calls to improperly treat many
requests as errors, displaying an error banner instead of calling the right
function. We now inspect the error and see if it has a proper JSON payload.
If it does, we call the "success" handler.

This restores our functionality, but is a temporary fix. Longer-term, we
should properly use the success/error callbacks in all cases.