Start implementing ReviewRequestDraftResource.
So far, all the field setting is there. The deprecated URLs still work for
that, and all the data is returned as expected.

The new PUT/POST handlers for this work about the same as the previous, but
now more fields are accepted, and some fields that were marked as valid before
but shouldn't have been are now invalid.

Error responses have been better standardized.