Reorganize and fix up the comment-related resources.
The diff and screenshot comment resources now have specific subclasses
for the various areas in the resource tree. They derive from base
resource classes for the particular comment type, for the common information.

For the old behavior of review comments, where both diff and screenshot
comments were returned, a special deprecated resource had to be created
that would combine the results of both lists.

The new and deprecated unit tests for these all now pass.