Factor in the starred state on a review request for caching.
If a user starred a review request and then reloaded, it would look like
the review request wasn't starred. It turned out we weren't paying attention
to this state when generating the ETag. Now we factor this in, so starring
will cause a different ETag.

Reviewed at http://reviews.reviewboard.org/r/1525/