Fix starring/unstarring of groups.
Starring/unstarring groups was broken in a few ways. Clicking the
star in the Groups pages would result in a JavaScript error about
$.fundQueue being defined. This was the result of including jquery.js
twice. The second inclusion overwrote the $ namespace.

We also weren't showing the star status for groups properly, due to
not adding the new starred group query aggregation.

Fixes bug #1538