Fix an off-by-one and rounding error with screenshot comments.
We had a few positioning issues with screenshot comments that would cause them
to be positioned incorrectly.

First, the math used to figure out the position of the comments would often be
given numbers like 200.5. We now make sure to round these down.

Second, we still had an off-by-one. A comment that was created at 0,0 would
be saved as being at 1,1.

We now make sure to account for this math in all cases.

Reviewed at http://reviews.reviewboard.org/r/1526/