Change ReviewRequestResource.create to take a 'repository' parameter.
Previously, both 'repository_id' and 'repository_path' were accepted, but
were mutually exclusive. Now a single 'repository' is accepted, and can be
either an ID or an actual path. If the value is numeric, then it's treated
as an ID, otherwise it's treated as a path.

This simplifies the logic and the requirements for the caller. It also fixes
some useless errors when using 'repository_id'.