Add proper loading and ready notification to RB.ReviewRequest.
This introduces a new ready() function on RB.ReviewRequest that works
similarly to jQuery's ready() function. Any call that operates on a
RB.ReviewRequest should wrap it in a review_request.ready() call. This will
guarantee that the state of the review request has been loaded. This may
include some variables, child hrefs, and more.

ready() will only perform a load of the resource the first time it's called.