Use CSS for the comment flags, make them multi-line, and improve positioning.
Previously, we used images for the comment flags. This looked really bad
when zoomed in. We also overlapped the line numbers.

Now we render the comment flags as little rounded rectangles with
shadows. The little pointer is gone, which is fine given that it's obvious
which line the flag is pointing to. We also leave enough room for the flag
so that the line number is not overlayed. This helps with seeing which
line a comment flag is on, and also helps with selecting lines that already
have a flag.

The comment flags were always shown only on one line, which didn't really
give a good idea as to how many lines the comment spanned. Now the comment
flag spans the number of lines it represents.

As a nice little touch, the comment flags are now slightly highlighted when
the starting line or the flag itself is selected.

Fixes bug #1131
Reviewed at http://reviews.reviewboard.org/r/1586/