Add a nice icon for iPad and iPhone bookmarks.
This introduces a nice default icon for the iPad and iPhone home screens.
It will show up when bookmarked to the home screen, and resembles the standard
Review Board logo except slighly stylized and rotated to match the standard
rounded icon frame of the rest of the icons on the home screen.