Switch to using python-memcached instead of cmemcache, and install by default.
cmemcache is deprecated, and on some systems was a pain to install anyway.
Django now recommends using python-memcached, a pure Python memcached client.
We now depend on this out of the box in order to simplify installation. The
upcoming 1.0.9 and 1.5 beta 3 versions will install this automatically, but
for now, there's documentation on installing it manually.

Fixes bug #1446