Fix rb-site to better handle empty directories.
If a user attempted to install a site using rb-site and ended up having
to start over, we would sometimes leave behind an empty directory. A second
attempt would tell the user that they needed to be an administrator, which
was misleading.

The error is now updated saying to also make sure that the user doesn't
have any files in that directory. We also attempt to remove the directory
first, if possible.

Fixes bug #1468
Reviewed at http://reviews.reviewboard.org/r/1600/