Update the database migration scripts to work better.
The old database migration scripts have been abandoned and weren't usable.
This updates them to work with the "new" directory structure and the models
we've added since deprecating them. With any luck, this should help people
who need to move from one database type to another.

In the future, we'll want to make this part of rb-site.