Support marking files as deleted in the diff viewer.
We've always treated deleted files as just another file change. A file with
a bunch of removes. However, this isn't as clear, and it's often just extra
noise. In the case of CVS, it actually doesn't work, because the files
referenced in the diff don't contain enough data to locate the source file.

Diff parsers can now flag a file as being deleted, which will prevent
any file existence checks and instead indicate in the diff viewer that the
file was deleted. This is similar in appearance to the binary indicator.

Right now, only CVS supports this, but support can easily be added to other
SCMTools later.

Fixes bug #394
Reviewed at http://reviews.reviewboard.org/r/1685/